requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
matplotlib==3.8.2
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
scikit-learn==1.3.2
Flask==3.0.0 
Flask-Cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0